+++ crates/wdl-format/src/v1/struct.rs
@@ -160,7 +160,7 @@ pub fn format_literal_struct(
         (&member).write(stream, config);
         if let Some(comma) = commas.next() {
             (comma).write(stream, config);
-        } else {
+        } else if config.trailing_commas {
             stream.push_literal(",".to_string(), SyntaxKind::Comma);
         }
         stream.end_line();